Extract server-state to break heavy import chain in bridge process
The stdio bridge imported readServerState from mcp-http-server.ts, which
transitively pulled in mcp-server.ts and the entire server stack (MCP SDK,
embeddings, NoteRepository, tool handlers). Bridge processes were ~110 MB
instead of the expected ~20 MB.

Extracted ServerState, readServerState, writeServerState, removeServerState
into a standalone server-state.ts that only depends on fs, path, logger,
config, and version. The bridge now imports from server-state.ts, breaking
the transitive dependency on the heavy server modules.
